+++ scripts/check-duplicates.mts
@@ -23,6 +23,7 @@ const targets = [
   "skills",
   "config",
   "node-runtime-update.mjs",
+  "node-sqlite.mjs",
   "node-version.mjs",
   "openclaw.mjs",
   "tsdown.ai.config.ts",
