+++ src/commands/node-daemon-runtime.test.ts
@@ -55,6 +55,7 @@ describe.skipIf(process.platform === "win32")("node-host runtime install boundar
       stdout: JSON.stringify({
         nodeVersion: "26.8.1",
         sqliteVersion: "3.53.4",
+        sqliteProbe: { available: true, version: "3.53.4", text: true, blob: true, json: true },
         nodeSharedSqlite: false,
       }),
       stderr: "",
