+++ test/scripts/install-ps1.test.ts
@@ -510,9 +510,8 @@ try {
           scriptWithoutEntryPoint,
           "function Get-Command { [pscustomobject]@{ Source = 'Invoke-FixtureNode' } }",
           "function Invoke-FixtureNode {",
-          "  param($argument)",
           "  $global:LASTEXITCODE = 0",
-          "  if ($argument -eq '-v') { return $script:FixtureVersion }",
+          "  if ($args[0] -eq '-v') { return $script:FixtureVersion }",
           "  $input | Out-Null",
           "  return $script:FixtureSqlite",
           "}",
@@ -653,7 +652,8 @@ try {
           "function private-node-fixture {",
           "  $global:LASTEXITCODE = 0",
           "  if ($args[0] -eq '-v') { return 'v26.1.0' }",
-          "  return $script:FixtureSqliteVersion",
+          "  $input | Out-Null",
+          "  return (@{ available = $true; version = $script:FixtureSqliteVersion; text = $true; blob = $true; json = $true } | ConvertTo-Json -Compress)",
           "}",
           "function Get-Command { throw 'unexpected ambient runtime lookup' }",
           "$cases = @{",
