+++ test/scripts/vitest-worker-artifacts.test.ts
@@ -1376,9 +1376,8 @@ export default class {
         fs.writeFileSync(
           dependency,
           `import { fixtureMajor } from "../../${privatePackage}/src/index.js";\n` +
-            fs
-              .readFileSync(dependency, "utf8")
-              .replace("major: 3, minor: 51", "major: fixtureMajor, minor: 51"),
+            'import { isSqliteWalResetSafeVersion as isSafeVersion } from "../../node-sqlite.mjs";\n' +
+            "export function isSqliteWalResetSafeVersion(value: string) { return fixtureMajor === 3 && isSafeVersion(value); }\n",
         );
         const compilerUrl = pathToFileURL(path.join(fixture, compilerModule)).href;
         const client = `
